Moved toasting.ts to utils\frontend
